達成済み任務を任務進捗表から消せるようにした
